iOS 5和modern ObjC syntax的坑真他妈大！把代码整个过一遍！

+++ SNBlockedCallHistoryViewController.h
@@ -9,7 +9,7 @@
 	NSMutableArray *timeArray;
 	NSMutableArray *readArray;
 	NSMutableSet *bulkSet;
-	int chosenRow;
+	NSUInteger chosenRow;
 }
 - (void)loadDatabaseSegment;
 - (void)selectAll:(UIBarButtonItem *)buttonItem;
