+++ SNPrivateCallHistoryViewController.h
@@ -8,7 +8,7 @@
 	NSMutableArray *contentArray;
 	NSMutableArray *timeArray;
 	NSMutableSet *bulkSet;
-	int chosenRow;
+	NSUInteger chosenRow;
 }
 - (void)loadDatabaseSegment;
 - (void)selectAll:(UIBarButtonItem *)buttonItem;
